go지갑 api 연동

+++ lib/platform/models/exchange_stik_price_model.dart
@@ -7,7 +7,7 @@ class ExchangeStikPriceModel {
   String? fromSymbol;
   double? fromUiAmount;
   String? toSymbol;
-  double? toUiAmount;
+  int? toUiAmount;
 
   ExchangeStikPriceModel({
     this.fromSymbol,
